process_manager/requirements.txt
redis==5.0.1
psutil==5.9.6
python-dotenv==1.0.0
dramatiq==1.15.0
pydantic==2.5.0
loguru==0.7.2